// frontend/src/components/home/FeaturedModels/FeaturedModels.types.ts
/**
 * ===============================================================
 * Author: ultramegared
 * Project: Magic Touch Designs
 * File: FeaturedModels.types.ts
 * Module: Home
 * Language: TypeScript
 * Description:
 * Featured Models component types.
 * ===============================================================
 */

export interface FeaturedModel {

    id: number;

    name: string;

    image: string;

    price: number;

}